packages/src/ocr_service/requirements.txt
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
pdf2image==1.17.0
